refactor(automation): decompose CIDriver into 4 SRP collaborators (combined #1179+#1289)
Decompose the CIDriver god-class (3,570 lines on current main) into four
narrow SRP collaborators wired by the Dependency Inversion Principle:

- pr_discovery.py         viewer-login caching + PR enumeration
- ci_check_inspector.py   required CI check queries (gh pr checks)
- ci_fix_orchestrator.py  CI fix session management (codex + claude)
- post_merge_processor.py drive-green /learn + /compact after merges

Each collaborator receives only the specific Callable providers it needs
(never self/CIDriver). All injected callables are lambda-wrapped so
patch.object(driver, "_method") intercepts through the lambda at call
time; every existing patch.object(CIDriver, "_method") target is
preserved via a delegation stub. ci_driver.py drops to 2,449 lines.

The corner-case fix-flow methods shipped after the two reference PRs are
preserved with exact behavior on CIDriver: _resolve_dirty_pr (#1347),
_resolve_blocked_pr / _address_threads_once (#1356), and the resolve_dirty
recursion guard on _recheck_and_arm_after_fix (#1355).

Supersedes #1292 and #1344, re-implemented against current main.
