+++ hephaestus/automation/ci_driver.py
@@ -18,7 +18,6 @@ import subprocess
 import threading
 import time
 from concurrent.futures import FIRST_COMPLETED, Future, ThreadPoolExecutor, wait
-from datetime import datetime, timezone
 from pathlib import Path
 from typing import Any
 
@@ -26,7 +25,6 @@ from hephaestus.agents.runtime import (
     add_agent_argument,
     is_codex,
     resolve_agent,
-    resume_codex_session,
     run_codex_session,
     session_agent_matches,
 )
@@ -40,48 +38,43 @@ from .address_review import (
 )
 from .advise_runner import run_advise
 from .arming_state import ArmingStateStore
+from .ci_check_inspector import FAILING_CHECK_CONCLUSIONS as FAILING_CHECK_CONCLUSIONS  # re-export
+from .ci_check_inspector import CICheckInspector
+from .ci_fix_orchestrator import CIFixOrchestrator
 from .claude_invoke import invoke_claude_with_session
-from .claude_models import advise_model, implementer_model
+from .claude_models import advise_model
 from .claude_timeouts import (
     advise_claude_timeout,
-    ci_driver_claude_timeout,
     ci_poll_max_wait,
-    learn_claude_timeout,
 )
 from .git_utils import (
-    get_repo_info,
     get_repo_root,
     get_repo_slug,
     issue_ref,
     pr_ref,
-    push_current_branch_with_lease_on_divergence,
-    rebase_worktree_onto,
-    run,
-    sync_worktree_to_remote_branch,
 )
 from .github_api import (
-    GitHubUnavailableError,
     _gh_call,
     gh_issue_json,
     gh_pr_checks,
     gh_pr_list_unresolved_threads,
     gh_pr_resolve_thread,
 )
-from .learn import build_learn_prompt, compact_session, mnemosyne_update_evidence
 from .models import CIDriverOptions, WorkerResult
+from .post_merge_processor import PostMergeProcessor
+from .pr_discovery import PRDiscovery
 from .pr_manager import pr_has_implementation_go_label
 from .prompts import get_advise_prompt_builder
-from .session_naming import AGENT_ADVISE, AGENT_CI_DRIVER
+from .session_naming import AGENT_ADVISE
 from .status_tracker import StatusTracker
 from .worktree_manager import WorktreeManager
 
 logger = logging.getLogger(__name__)
 
-# Conclusion values that indicate a PR's check rollup is failing in a way
-# drive-green can act on. SUCCESS / SKIPPED / NEUTRAL / PENDING are
-# explicitly excluded. Shared with loop_runner._count_failing_prs so the
-# SKIP gate and the actual work list never drift (#819).
-FAILING_CHECK_CONCLUSIONS: frozenset[str] = frozenset({"FAILURE", "CANCELLED", "TIMED_OUT"})
+# FAILING_CHECK_CONCLUSIONS moved to ci_check_inspector.py (#1357) and
+# re-exported from the import block above for backward compatibility — the
+# module-level _pr_is_failing predicate and loop_runner._count_failing_prs
+# both still import it from here.
 
 # Max address-review passes for a green-but-BLOCKED PR before leaving it armed
 # (#1348). The progress guard stops earlier whenever a pass resolves no new
@@ -146,11 +139,51 @@ class CIDriver:
         # arming record (and therefore its own /learn capture once the PR
         # actually merges in a subsequent run). #840 +on top of #834.
         self.shared_pr_issues: dict[int, list[int]] = {}
-        # Viewer login cache (#821). Resolved lazily on first use of the
-        # author filter; empty string means "not yet resolved". When
-        # options.include_all_authors=True the filter is bypassed and this
-        # stays empty so a broken `gh auth` does not block --all runs.
-        self._viewer_login: str = ""
+
+        # Narrow-callable collaborators extracted from CIDriver (#1357 /
+        # refs #1179, #1289). Each receives only the specific Callable[[], T]
+        # providers / bound methods it needs (DIP) — never ``self``. The
+        # injected callables are wrapped in lambdas (or passed as bound methods
+        # that delegate further) so ``patch.object(driver, "_method")`` in tests
+        # still intercepts through the indirection at call time. The viewer-login
+        # cache (#821) now lives inside PRDiscovery.
+        self._pr_discovery = PRDiscovery(
+            options_provider=lambda: self.options,
+            status_tracker_provider=lambda: self.status_tracker,
+            repo_root_provider=lambda: self.repo_root,
+            pr_merge_state_provider=lambda pr_number: self._pr_merge_state(pr_number),
+        )
+        self._check_inspector = CICheckInspector(
+            get_pr_branch=lambda pr_number: self._get_pr_branch(pr_number),
+            options_provider=lambda: self.options,
+        )
+        self._fix_orchestrator = CIFixOrchestrator(
+            options_provider=lambda: self.options,
+            repo_root_provider=lambda: self.repo_root,
+            state_dir_provider=lambda: self.state_dir,
+            status_tracker_provider=lambda: self.status_tracker,
+            get_pr_branch=lambda pr_number: self._get_pr_branch(pr_number),
+            get_worktree_path=lambda issue_number, pr_number: self._get_worktree_path(
+                issue_number, pr_number
+            ),
+            format_review_threads_block=lambda pr_number: self._format_review_threads_block(
+                pr_number
+            ),
+            failing_required_check_names=lambda pr_number: self._failing_required_check_names(
+                pr_number
+            ),
+        )
+        self._post_merge = PostMergeProcessor(
+            options_provider=lambda: self.options,
+            repo_root_provider=lambda: self.repo_root,
+            get_worktree_path=lambda issue_number, pr_number: self._get_worktree_path(
+                issue_number, pr_number
+            ),
+            load_arming_state=lambda issue_number: self._load_arming_state(issue_number),
+            save_arming_state=lambda issue_number, record: self._save_arming_state(
+                issue_number, record
+            ),
+        )
 
     def run(self) -> dict[int, WorkerResult]:  # noqa: C901  # orchestration: thread pool + finally + preserve report across exception paths
         """Run the CI driver on all configured issues.
@@ -300,140 +333,12 @@ class CIDriver:
         return results
 
     def _list_open_prs_remaining(self) -> list[dict[str, Any]]:
-        """Return the list of open PRs left on the repo after the drive (#838).
-
-        A repo is only truly "driven" when there are zero open PRs left. The
-        per-issue ``_drive_issue`` loop's notion of success — every issue's
-        PR moved to green and/or got auto-merge enabled — does NOT imply the
-        repo is clean: PRs that have not yet merged (auto-merge waiting on
-        CI), PRs from issues outside the input set, and PRs opened by
-        humans/other-automation all leave open work behind.
-
-        Uses ``gh api --paginate`` so the result is the FULL set of open PRs,
-        not a capped prefix. A repo with hundreds of dependabot PRs would
-        otherwise pass the done-check after looking at only 100 of them.
-
-        Returns:
-            One dict per open PR with keys ``number``, ``title``,
-            ``headRefName``, ``autoMergeRequest`` (None or the auto-merge
-            metadata blob), and ``mergeStateStatus`` / ``mergeable`` (the
-            per-PR merge-state, fetched separately because the REST list
-            endpoint does not populate ``mergeable`` reliably — see #1328).
-            Empty list iff the repo is clean.
-
-        """
-        try:
-            owner, repo = get_repo_info(self.repo_root)
-        except RuntimeError as exc:
-            logger.error("Could not resolve repo owner/name to list open PRs: %s", exc)
-            # Unknown ownership ⇒ treat as not-done so operators investigate.
-            return [{"number": -1, "title": "(unknown: cannot resolve repo)"}]
-
-        # ``gh api --paginate`` walks ``Link: rel="next"`` headers and emits
-        # a single concatenated JSON array across all pages. ``per_page=100``
-        # is GitHub's max page size; we issue the minimum number of calls.
-        # We use ``gh api`` directly (not ``gh pr list``) because the latter
-        # caps at ``--limit`` even with paginate semantics; gh's REST proxy
-        # paginates without an upper bound.
-        try:
-            result = _gh_call(
-                [
-                    "api",
-                    "--paginate",
-                    f"/repos/{owner}/{repo}/pulls?state=open&per_page=100",
-                ],
-                check=False,
-            )
-            raw_pulls: list[dict[str, Any]] = json.loads(result.stdout or "[]")
-        except (subprocess.CalledProcessError, json.JSONDecodeError) as exc:
-            # If we cannot determine the open-PR count, the safest default is
-            # to assume the repo is NOT done — surface the unknown state as a
-            # failure so operators don't walk away on a false-green.
-            logger.error("Could not list open PRs to verify repo done-state: %s", exc)
-            return [{"number": -1, "title": "(unknown: gh api pulls failed)"}]
-
-        # The REST shape exposes ``head.ref`` and ``auto_merge`` (snake_case);
-        # normalise to the gh-CLI shape consumers downstream already use.
-        viewer = "" if self.options.include_all_authors else self._resolve_viewer_login()
-        normalised: list[dict[str, Any]] = []
-        for pr in raw_pulls:
-            user = pr.get("user") or {}
-            if viewer and user.get("login") != viewer:
-                if user.get("login") is None:
-                    logger.warning(
-                        "PR #%s has no user.login; skipping under author filter (#821)",
-                        pr.get("number"),
-                        extra={
-                            "missing_field": "user.login",
-                            "filter": "author",
-                            "pr_number": pr.get("number"),
-                        },
-                    )
-                continue  # #821: hide other-author PRs from the done-gate sweep
-            labels = pr.get("labels") or []
-            number = pr.get("number")
-            merge_state, mergeable = self._pr_merge_state(number)
-            normalised.append(
-                {
-                    "number": number,
-                    "title": pr.get("title", ""),
-                    "headRefName": (pr.get("head") or {}).get("ref", ""),
-                    "autoMergeRequest": pr.get("auto_merge"),
-                    "mergeStateStatus": merge_state,
-                    "mergeable": mergeable,
-                    "labels": [
-                        label.get("name", "") for label in labels if isinstance(label, dict)
-                    ],
-                    "isBot": user.get("type") == "Bot",
-                }
-            )
-        return normalised
+        """Delegate to PRDiscovery (extracted #1357)."""
+        return self._pr_discovery.list_open_prs_remaining()
 
     def _pr_merge_state(self, pr_number: Any) -> tuple[str, str]:
-        """Return ``(mergeStateStatus, mergeable)`` for a single PR (#1328).
-
-        The REST ``/pulls`` list endpoint does NOT populate ``mergeable`` /
-        ``mergeable_state`` reliably (GitHub computes the merge-state lazily and
-        omits it from list responses), so the done-gate cannot tell a
-        permanently-CONFLICTING armed PR apart from one that is genuinely still
-        merging. A per-PR ``gh pr view`` forces GitHub to compute the merge
-        state, matching how the rest of this module queries merge-state
-        (``_attempt_mechanical_rebase`` / ``_gh_pr_state``).
-
-        Args:
-            pr_number: PR number to query.
-
-        Returns:
-            ``(mergeStateStatus, mergeable)`` upper-cased. Empty strings when the
-            number is the unknown-marker sentinel or the query fails — an unknown
-            merge-state must never be misread as CONFLICTING.
-
-        """
-        if not isinstance(pr_number, int) or pr_number < 0:
-            return "", ""
-        try:
-            result = _gh_call(
-                [
-                    "pr",
-                    "view",
-                    str(pr_number),
-                    "--json",
-                    "mergeStateStatus,mergeable",
-                ],
-                check=False,
-            )
-            state = dict(json.loads(result.stdout or "{}"))
-        except (subprocess.CalledProcessError, json.JSONDecodeError) as exc:
-            logger.warning(
-                "Could not fetch PR #%s merge-state for done-gate; treating as unknown: %s",
-                pr_number,
-                exc,
-            )
-            return "", ""
-        return (
-            str(state.get("mergeStateStatus") or "").upper(),
-            str(state.get("mergeable") or "").upper(),
-        )
+        """Delegate to PRDiscovery (extracted #1357)."""
+        return self._pr_discovery.pr_merge_state(pr_number)
 
     def _arm_all_unarmed_open_prs(self, open_prs: list[dict[str, Any]]) -> list[dict[str, Any]]:
         """Arm auto-merge on every implementation-GO un-armed open PR (#882).
@@ -473,201 +378,20 @@ class CIDriver:
         return self._list_open_prs_remaining()
 
     def _resolve_viewer_login(self) -> str:
-        """Return the authenticated `gh api user` login. Fail CLOSED on error.
-
-        Lazy + cached: only called when the author filter is active. Raises
-        ``RuntimeError`` with operator guidance on any failure so a broken
-        `gh` auth never silently widens scope to all PRs (#821 POLA).
-        """
-        if self._viewer_login:
-            return self._viewer_login
-        try:
-            result = _gh_call(["api", "user", "-q", ".login"], check=True)
-        except (
-            subprocess.CalledProcessError,
-            FileNotFoundError,
-            GitHubUnavailableError,
-        ) as exc:
-            raise RuntimeError(
-                "Could not resolve viewer login via `gh api user`: "
-                f"{exc}. Re-authenticate with `gh auth login`, or pass "
-                "--all to opt out of the @me filter (#821)."
-            ) from exc
-        login = (result.stdout or "").strip()
-        if not login:
-            raise RuntimeError(
-                "Could not resolve viewer login via `gh api user`: "
-                "empty response. Re-authenticate with `gh auth login`, "
-                "or pass --all to opt out of the @me filter (#821)."
-            )
-        self._viewer_login = login
-        return login
+        """Delegate to PRDiscovery (extracted #1357)."""
+        return self._pr_discovery.resolve_viewer_login()
 
     def _discover_bot_prs(self) -> dict[int, int]:
-        """Enumerate every open ``is_bot=true`` PR on the repo (#848).
-
-        Bot PRs (Dependabot, github-actions, etc.) carry NO ``Closes #N``
-        link to an issue, so the issue-driven discovery path can never see
-        them — they are architecturally invisible. Without this enumeration
-        a repo can sit with dozens of stranded Dependabot PRs forever while
-        the ecosystem script cheerfully reports "driven" because every
-        listed issue had no matching PR.
-
-        Returns a mapping where each bot PR's number is used both as the
-        synthetic issue key AND the PR number. Downstream code is taught
-        (``_is_bot_pr_mode``) to detect the equality and skip issue-data
-        fetches that would 404 on a synthetic key.
-
-        Returns:
-            Mapping of ``pr_number -> pr_number`` for every open bot PR.
-            Empty dict if the ``gh api`` pulls lookup fails or returns
-            nothing — bot discovery must never abort the drive on a list
-            failure.
-
-        Raises:
-            RuntimeError: When the default @me author filter is active
-                (``--all`` not set) and viewer-login resolution fails. This
-                fail-CLOSED abort is intentional per #821 (POLA): a broken
-                ``gh auth`` must never silently widen scope to every author's
-                PRs. Pass ``--all`` to opt out of the filter and bypass the
-                resolver entirely.
-
-        """
-        try:
-            owner, repo = get_repo_info(self.repo_root)
-        except RuntimeError as exc:
-            logger.info("Bot-PR discovery skipped: could not resolve owner/name (%s)", exc)
-            return {}
-
-        try:
-            result = _gh_call(
-                [
-                    "api",
-                    "--paginate",
-                    f"/repos/{owner}/{repo}/pulls?state=open&per_page=100",
-                ],
-                check=False,
-            )
-            raw_pulls: list[dict[str, Any]] = json.loads(result.stdout or "[]")
-        except (
-            subprocess.CalledProcessError,
-            subprocess.TimeoutExpired,
-            OSError,
-            json.JSONDecodeError,
-        ) as exc:
-            logger.info("Bot-PR discovery skipped: gh api failed (%s)", exc)
-            return {}
-
-        viewer = "" if self.options.include_all_authors else self._resolve_viewer_login()
-        bot_prs: dict[int, int] = {}
-        for pr in raw_pulls:
-            user = pr.get("user") or {}
-            if user.get("type") != "Bot":
-                continue
-            if viewer and user.get("login") != viewer:
-                if user.get("login") is None:
-                    logger.warning(
-                        "PR #%s has no user.login; skipping under author filter (#821)",
-                        pr.get("number"),
-                        extra={
-                            "missing_field": "user.login",
-                            "filter": "author",
-                            "pr_number": pr.get("number"),
-                        },
-                    )
-                continue  # #821: not viewer-owned and --all not set
-            number = pr.get("number")
-            if isinstance(number, int):
-                bot_prs[number] = number
-
-        if bot_prs:
-            logger.info(
-                "Discovered %s open bot-authored PR(s): %s",
-                len(bot_prs),
-                sorted(bot_prs),
-            )
-        return bot_prs
+        """Delegate to PRDiscovery (extracted #1357)."""
+        return self._pr_discovery.discover_bot_prs()
 
     def _discover_failing_prs(self) -> dict[int, int]:
-        """Enumerate open non-draft PRs whose checks failed or merge is BLOCKED.
-
-        Symmetrical to ``_discover_bot_prs``: the issue→PR direction (Closes #N)
-        misses every PR with no Closes line and every PR linked to a closed
-        issue (issue body §1, §2). One CLI call, PR-keyed, synthetic-issue
-        invariant (pr_number == issue_number) so downstream ``_is_bot_pr_mode``
-        short-circuits ``gh issue view`` identically to the bot path.
-
-        Bounded by gh's --limit 1000 (its documented hard upper). A repo with
-        more than 1000 failing open PRs is pathological — we log a WARNING
-        so operators see the truncation rather than silently dropping work.
-
-        Returns:
-            Mapping pr_number -> pr_number for every failing open PR.
-            Empty dict on any lookup failure — discovery must never abort
-            the drive.
-
-        """
-        try:
-            owner, repo = get_repo_info(self.repo_root)
-        except RuntimeError as exc:
-            logger.info("Failing-PR discovery skipped: could not resolve owner/name (%s)", exc)
-            return {}
-        try:
-            result = _gh_call(
-                [
-                    "pr",
-                    "list",
-                    "--repo",
-                    f"{owner}/{repo}",
-                    "--state",
-                    "open",
-                    "--limit",
-                    "1000",
-                    "--json",
-                    "number,isDraft,statusCheckRollup,mergeStateStatus",
-                ],
-            )
-            pulls: list[dict[str, Any]] = json.loads(result.stdout or "[]")
-        except (
-            subprocess.CalledProcessError,
-            subprocess.TimeoutExpired,
-            OSError,
-            json.JSONDecodeError,
-        ) as exc:
-            logger.info("Failing-PR discovery skipped: gh pr list failed (%s)", exc)
-            return {}
-        if len(pulls) >= 1000:
-            logger.warning(
-                "Failing-PR discovery hit gh's 1000-PR cap on %s/%s — "
-                "additional failing PRs may exist and are not visible to this run.",
-                owner,
-                repo,
-            )
-        failing: dict[int, int] = {}
-        for pr in pulls:
-            number = pr.get("number")
-            if not isinstance(number, int):
-                continue
-            if _pr_is_failing(pr):
-                failing[number] = number
-        if failing:
-            logger.info(
-                "Discovered %s open failing PR(s): %s",
-                len(failing),
-                sorted(failing),
-            )
-        return failing
+        """Delegate to PRDiscovery (extracted #1357)."""
+        return self._pr_discovery.discover_failing_prs(_pr_is_failing)
 
     def _is_bot_pr_mode(self, issue_number: int, pr_number: int) -> bool:
-        """Return True iff this work item is a synthetic-issue bot PR (#848).
-
-        The bot-PR enumeration uses the PR number as a stand-in for an
-        issue number because Dependabot PRs have no associated issue.
-        Anywhere we would normally call ``gh issue view <issue_number>``
-        we must instead short-circuit; this helper centralises the check
-        so a single rule (issue == pr) keeps both ends honest.
-        """
-        return issue_number == pr_number
+        """Delegate to PRDiscovery (extracted #1357)."""
+        return self._pr_discovery.is_bot_pr_mode(issue_number, pr_number)
 
     def _discover_prs(self, issue_numbers: list[int]) -> dict[int, int]:  # noqa: C901  # orchestration: multi-PR dedup with fallback discovery across issue/PR mappings
         """Pre-discover open PRs for all issues, deduped by PR.
@@ -1005,122 +729,10 @@ class CIDriver:
         pr_number: int,
         acquired_slot: int,
     ) -> bool:
-        """Rebase a behind/conflicting PR onto its base branch with no agent (#871).
-
-        The cheap, deterministic path: a PR that is merely behind its base (or
-        whose changes don't textually overlap) is rebased and pushed here. Only a
-        PR whose rebase hits real conflicts falls through to the CI-fix agent.
-
-        Flow:
-
-        1. Query ``mergeStateStatus`` / ``mergeable`` / ``baseRefName``. Skip
-           (return ``False``) unless the PR is ``BEHIND`` or ``DIRTY`` /
-           ``CONFLICTING`` — a PR already on top of its base needs no rebase and
-           the normal check-status path handles it.
-        2. Sync the worktree to the PR head, then ``rebase_worktree_onto`` the
-           base branch.
-        3. Clean rebase → push ``HEAD:<pr-head>`` with ``--force-with-lease`` and
-           return ``True``. The push re-triggers CI; the caller's poll loop
-           evaluates the rebased head.
-        4. Conflicts → the rebase is aborted inside ``rebase_worktree_onto``;
-           return ``False`` so the caller continues to the agent path.
-
-        Any unexpected git/subprocess error is logged and swallowed (returns
-        ``False``) — a mechanical-rebase failure must never crash the worker; the
-        agent path is always the safe fallback.
-
-        Args:
-            issue_number: GitHub issue number for the PR.
-            pr_number: PR number to rebase.
-            acquired_slot: Worker slot ID for status tracking.
-
-        Returns:
-            ``True`` if the PR was mechanically rebased and pushed; ``False`` if
-            no rebase was needed, the rebase conflicted, or an error occurred.
-
-        """
-        try:
-            result = _gh_call(
-                [
-                    "pr",
-                    "view",
-                    str(pr_number),
-                    "--json",
-                    "mergeStateStatus,mergeable,headRefName,baseRefName",
-                ],
-                check=False,
-            )
-            state = dict(json.loads(result.stdout or "{}"))
-        except (subprocess.CalledProcessError, json.JSONDecodeError) as exc:
-            logger.warning(
-                "Issue #%s: could not fetch PR #%s merge state for rebase; "
-                "skipping mechanical rebase: %s",
-                issue_number,
-                pr_number,
-                exc,
-            )
-            return False
-
-        merge_state = str(state.get("mergeStateStatus") or "").upper()
-        # Only behind-base / conflicting PRs need a rebase. CLEAN / BLOCKED /
-        # UNSTABLE / HAS_HOOKS PRs are already on top of their base — let the
-        # check-status path handle them. (BLOCKED is the review-gated case.)
-        if merge_state not in ("BEHIND", "DIRTY", "CONFLICTING"):
-            return False
-
-        pr_head_branch = str(state.get("headRefName") or "") or self._get_pr_branch(pr_number)
-        base_branch = str(state.get("baseRefName") or "main") or "main"
-        if not pr_head_branch:
-            logger.warning(
-                "Issue #%s: PR #%s has no resolvable head branch; skipping rebase",
-                issue_number,
-                pr_number,
-            )
-            return False
-
-        self.status_tracker.update_slot(
-            acquired_slot,
-            f"{issue_ref(issue_number)}: mechanical rebase onto {base_branch}",
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.attempt_mechanical_rebase(
+            issue_number, pr_number, acquired_slot
         )
-
-        try:
-            worktree_path = self._get_worktree_path(issue_number, pr_number)
-            # Land on the PR's actual remote head before rebasing so we replay the
-            # PR's commits (not a stale local ref) onto the base (#832).
-            sync_worktree_to_remote_branch(worktree_path, pr_head_branch)
-
-            if not rebase_worktree_onto(worktree_path, base_branch):
-                logger.info(
-                    "Issue #%s: PR #%s (%s) has rebase conflicts onto %s; deferring to agent",
-                    issue_number,
-                    pr_number,
-                    merge_state,
-                    base_branch,
-                )
-                return False
-
-            # Clean rebase rewrote history — lease-push to the PR head. The helper
-            # no-ops cleanly if the rebase was already up to date (HEAD unchanged).
-            push_current_branch_with_lease_on_divergence(
-                worktree_path,
-                branch=pr_head_branch,
-                push_ref=f"HEAD:{pr_head_branch}",
-            )
-            logger.info(
-                "Issue #%s: mechanically rebased PR #%s onto %s and pushed (no agent)",
-                issue_number,
-                pr_number,
-                base_branch,
-            )
-            return True
-        except subprocess.CalledProcessError as exc:
-            logger.warning(
-                "Issue #%s: mechanical rebase of PR #%s failed (%s); falling through to agent",
-                issue_number,
-                pr_number,
-                (exc.stderr or exc.stdout or "")[:300],
-            )
-            return False
 
     def _run_advise(self, issue_number: int) -> str:
         """Pull prior learnings from ProjectMnemosyne before the CI fix loop.
@@ -1648,60 +1260,8 @@ class CIDriver:
         return self.worktree_manager.create_worktree(issue_number, branch)
 
     def _get_failing_ci_logs(self, pr_number: int) -> str:
-        """Fetch combined failure logs for recent failed CI runs on a PR.
-
-        Scopes the ``gh run list`` query to the PR's head branch so we only
-        see runs that belong to this PR rather than the most-recent repo-wide
-        runs (the previous repo-wide query could return runs for other PRs
-        and even other branches, making the logs useless for fixing *this* PR).
-
-        Args:
-            pr_number: GitHub PR number.
-
-        Returns:
-            Combined log string, truncated to 10 000 characters.
-
-        """
-        try:
-            branch = self._get_pr_branch(pr_number)
-            result2 = _gh_call(
-                [
-                    "run",
-                    "list",
-                    "--branch",
-                    branch,
-                    "--status",
-                    "failure",
-                    "--limit",
-                    "10",
-                    "--json",
-                    "databaseId,conclusion,name,headSha",
-                ],
-                check=False,
-            )
-            runs: list[dict[str, Any]] = json.loads(result2.stdout or "[]")
-            failed_runs = [r for r in runs if r.get("conclusion") == "failure"][:3]
-
-            logs: list[str] = []
-            for run_info in failed_runs:
-                run_id = run_info.get("databaseId")
-                run_name = run_info.get("name", str(run_id))
-                if not run_id:
-                    continue
-                try:
-                    log_result = _gh_call(
-                        ["run", "view", str(run_id), "--log-failed"],
-                        check=False,
-                    )
-                    logs.append(f"=== {run_name} ===\n{log_result.stdout[:3000]}")
-                except Exception as log_err:
-                    logger.debug("Could not fetch log for run %s: %s", run_id, log_err)
-
-            return "\n\n".join(logs)[:10000]
-
-        except Exception as e:
-            logger.warning("Could not fetch CI logs for PR #%s: %s", pr_number, e)
-            return ""
+        """Delegate to CICheckInspector (extracted #1357)."""
+        return self._check_inspector.get_failing_ci_logs(pr_number)
 
     # ------------------------------------------------------------------
     # Drive-green arming records (#840)
@@ -1742,33 +1302,8 @@ class CIDriver:
         *,
         succeeded: bool,
     ) -> None:
-        """Persist an explicit attempted/succeeded/failed learn result.
-
-        ``learn_captured_at`` is retained as the success timestamp for older
-        readers, but failure now records ``learn_status=failed`` without
-        claiming Mnemosyne was updated.
-        """
-        timestamp = time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime())
-        record["learn_attempted_at"] = timestamp
-        if succeeded:
-            record["learn_status"] = "succeeded"
-            record["learn_succeeded_at"] = timestamp
-            record["learn_captured_at"] = timestamp
-            record.update(
-                getattr(self, "_last_drive_green_learn_evidence", mnemosyne_update_evidence(""))
-            )
-        else:
-            record["learn_status"] = "failed"
-            record["learn_succeeded_at"] = None
-            record["learn_captured_at"] = None
-            record.update(
-                {
-                    "mnemosyne_update_status": "failed",
-                    "mnemosyne_update_urls": [],
-                    "mnemosyne_update_pr_numbers": [],
-                }
-            )
-        self._save_arming_state(issue_number, record)
+        """Delegate to PostMergeProcessor (extracted #1357)."""
+        self._post_merge.mark_drive_green_learn_result(issue_number, record, succeeded=succeeded)
 
     def _arm_drive_green(self, pr_number: int, pr_head_branch: str, pr_head_sha: str) -> None:
         """Record arming for every issue that resolved to ``pr_number``.
@@ -2302,72 +1837,16 @@ class CIDriver:
         return resolved
 
     def _failing_required_check_names(self, pr_number: int) -> list[str]:
-        """Return names of required checks that are currently failing.
-
-        Used by the no-commit retry path (#846) to name the actual offenders
-        verbatim in the force-engagement prompt. Returns an empty list if
-        the lookup fails — the caller treats that as "cannot prove still
-        red" and skips the retry rather than launching Claude blind.
-        """
-        try:
-            checks = gh_pr_checks(pr_number, dry_run=self.options.dry_run)
-        except Exception as exc:
-            logger.info(
-                "PR #%s: failed to re-check CI for no-commit retry decision (%s)",
-                pr_number,
-                exc,
-            )
-            return []
-        if not checks:
-            return []
-        required = [c for c in checks if c.get("required")] or checks
-        return [
-            c.get("name", "")
-            for c in required
-            if c.get("status") == "completed" and c.get("conclusion") == "failure"
-        ]
+        """Delegate to CICheckInspector (extracted #1357)."""
+        return self._check_inspector.failing_required_check_names(pr_number)
 
     def _tracked_worktree_changes(self, worktree_path: Path, issue_number: int) -> list[str]:
-        """Return tracked dirty status lines for a post-agent worktree.
-
-        Untracked tool output such as a local ``uv.lock`` is intentionally
-        ignored. A no-commit turn that left tracked files modified is still
-        actionable even when the remote has no red required checks, as happens
-        for merge-conflict/behind-branch repairs where the agent fixed files but
-        forgot the signed commit.
-        """
-        status = self._git_stdout_for_push_guard(
-            worktree_path,
-            issue_number,
-            ["git", "status", "--porcelain"],
-            "failed to inspect worktree status for no-commit retry",
-        )
-        if status is None:
-            return []
-        return [line for line in status.splitlines() if line.strip() and not line.startswith("?? ")]
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator._tracked_worktree_changes(worktree_path, issue_number)
 
     def _pending_required_check_names(self, pr_number: int) -> list[str]:
-        """Return names of required checks that are still in flight (not completed).
-
-        Used by the BLOCKED early-exit guard in ``_wait_for_pr_terminal`` to
-        distinguish branch-protection blocks (all checks green but conversations
-        unresolved) from pending-CI blocks (checks still running).  Returns an
-        empty list on lookup failure — the caller then conservatively assumes no
-        checks are pending and exits the poll.
-        """
-        try:
-            checks = gh_pr_checks(pr_number, dry_run=self.options.dry_run)
-        except Exception as exc:
-            logger.info(
-                "PR #%s: failed to fetch CI checks for BLOCKED pending guard (%s)",
-                pr_number,
-                exc,
-            )
-            return []
-        if not checks:
-            return []
-        required = [c for c in checks if c.get("required")] or checks
-        return [c.get("name", "") for c in required if c.get("status") != "completed"]
+        """Delegate to CICheckInspector (extracted #1357)."""
+        return self._check_inspector.pending_required_check_names(pr_number)
 
     def _force_engagement_prompt(
         self,
@@ -2380,66 +1859,15 @@ class CIDriver:
         review_threads_block: str,
         dirty_tracked_changes: list[str] | None = None,
     ) -> str:
-        """Build the retry prompt when the agent returned without committing (#846).
-
-        The retry must engage the agent enough to either (a) produce a real
-        fix or (b) explicitly say why CI cannot pass / merge. The prompt names
-        the failing checks and/or dirty tracked files verbatim, re-emphasises
-        the existing PR/branch invariant, and re-emphasises signed commits — a
-        no-commit retry is a contract violation that the agent has to address
-        head-on.
-        """
-        failing_block = "\n".join(f"- {n}" for n in failing_check_names) or "- (unknown)"
-        dirty_lines = dirty_tracked_changes or []
-        dirty_block = "\n".join(f"- {line}" for line in dirty_lines)
-        if dirty_block:
-            dirty_block = (
-                "\n\nThe local worktree also contains uncommitted tracked changes "
-                "from the previous turn. Review this existing diff first and either "
-                f"commit it after verification or amend it before committing:\n\n{dirty_block}\n"
-            )
-        remote_block = (
-            "The required CI checks below are STILL failing on the remote"
-            if failing_check_names
-            else "The remote checks may be green, but the PR still needs a committed "
-            "repair before the driver can push"
-        )
-        return (
-            f"{review_threads_block}"
-            f"## Force-Engagement Retry — Previous Turn Produced No Commit\n\n"
-            f"You just returned from a CI-fix session for PR {pr_ref(pr_number)} "
-            f"(issue {issue_ref(issue_number)}) WITHOUT producing a new commit on "
-            f"branch `{pr_head_branch}`. {remote_block}:\n\n"
-            f"{failing_block}\n\n"
-            f"{dirty_block}"
-            f"Returning no commit when required checks are still red is itself a "
-            f"bug; returning no commit after editing tracked files is also a bug. "
-            f"Fix the code so the failing checks pass and the PR can merge. If no "
-            f"code fix is possible, DO NOT commit a 'blocker' file: a new "
-            f"Markdown/docs file will itself fail the repo's lint gates (e.g. "
-            f"markdownlint) and turn one blocker into two. Instead leave the tree "
-            f"unchanged and report the blocker via the `BLOCKED:` line below — do "
-            f"NOT commit any file to document it.\n\n"
-            f"Working directory: {worktree_path}\n"
-            f"Current branch (DO NOT change, DO NOT create a new branch): "
-            f"{pr_head_branch}\n\n"
-            f"Required behaviour:\n"
-            f"1. Re-read the failing check logs for the names listed above.\n"
-            f"2. Make the minimal change that addresses each failure.\n"
-            f"3. Run `pixi run python -m pytest tests/ -v` and "
-            f"`pre-commit run --all-files` locally to verify before committing. "
-            f"This MUST include any markdown/lint hooks — every file you add or "
-            f"edit has to pass the repo's own linters, with no rule disabled.\n"
-            f"4. **Every commit MUST be cryptographically signed (`git commit -S`).** "
-            f"NEVER use `--no-verify`. The repository's CI gate rejects unsigned "
-            f"commits and any commit that bypassed pre-commit hooks.\n"
-            f"5. Do NOT run `git checkout -b`, `git switch -c`, or any command "
-            f"that creates or switches branches — the fix has to land on "
-            f"`{pr_head_branch}`.\n"
-            f"6. Do NOT add a new top-level `CI_BLOCKER.md` or similar doc file to "
-            f"record the blocker — use the `BLOCKED:` line instead.\n\n"
-            f"If after the steps above you still cannot produce a commit, reply "
-            f"with a single line `BLOCKED: <one-sentence reason>` and stop."
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.force_engagement_prompt(
+            issue_number=issue_number,
+            pr_number=pr_number,
+            worktree_path=worktree_path,
+            pr_head_branch=pr_head_branch,
+            failing_check_names=failing_check_names,
+            review_threads_block=review_threads_block,
+            dirty_tracked_changes=dirty_tracked_changes,
         )
 
     def _record_repeated_no_commit(
@@ -2450,36 +1878,13 @@ class CIDriver:
         pr_head_branch: str,
         failing_check_names: list[str],
     ) -> None:
-        """Persist a marker for the next ecosystem run (#846).
-
-        Writes ``state_dir / "repeated-no-commit-<pr>.json"`` so a future
-        run (and the human reading the logs) can see which PRs got stuck
-        in the no-commit loop. We deliberately do NOT delete the arming
-        record here — the PR is still open and may yet land via another
-        actor; the marker file is purely a forensics aid.
-        """
-        marker = self.state_dir / f"repeated-no-commit-{pr_number}.json"
-        try:
-            marker.write_text(
-                json.dumps(
-                    {
-                        "issue_number": issue_number,
-                        "pr_number": pr_number,
-                        "pr_head_branch": pr_head_branch,
-                        "failing_required_checks": failing_check_names,
-                        "recorded_at": datetime.now(timezone.utc).isoformat(),
-                    },
-                    indent=2,
-                )
-                + "\n"
-            )
-        except OSError as exc:
-            logger.warning(
-                "Issue #%s: failed to write repeated-no-commit marker for PR #%s: %s",
-                issue_number,
-                pr_number,
-                exc,
-            )
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        self._fix_orchestrator.record_repeated_no_commit(
+            issue_number=issue_number,
+            pr_number=pr_number,
+            pr_head_branch=pr_head_branch,
+            failing_check_names=failing_check_names,
+        )
 
     def _invoke_agent_session(
         self,
@@ -2490,94 +1895,14 @@ class CIDriver:
         issue_number: int,
         pr_number: int,
     ) -> subprocess.CompletedProcess[str]:
-        """Dispatch a prompt to the configured agent (codex or claude).
-
-        Returns a CompletedProcess whose returncode signals success or failure:
-        - returncode == 0: agent ran successfully (stdout has output)
-        - returncode != 0: agent failed (stderr has details)
-
-        For codex, returncode is synthetic — AgentRunResult has no returncode
-        field; success is "no CalledProcessError" (hephaestus/agents/runtime.py).
-        For claude, returncode comes from the subprocess exit code.
-
-        CalledProcessError is absorbed from all codex paths into a non-zero
-        CompletedProcess. TimeoutExpired propagates to the caller so it can
-        log a distinct timeout message.
-        """
-        if is_codex(self.options.agent):
-            if session_id:
-                try:
-                    result = resume_codex_session(
-                        session_id,
-                        prompt,
-                        cwd=worktree_path,
-                        timeout=ci_driver_claude_timeout(),
-                    )
-                except subprocess.CalledProcessError as exc:
-                    logger.warning(
-                        "Issue #%s: Codex resume session %r failed for PR #%s; "
-                        "falling back to fresh session: %s",
-                        issue_number,
-                        session_id,
-                        pr_number,
-                        (exc.stderr or exc.stdout or "")[:300],
-                    )
-                    try:
-                        result = run_codex_session(
-                            prompt,
-                            cwd=worktree_path,
-                            timeout=ci_driver_claude_timeout(),
-                            sandbox="workspace-write",
-                        )
-                    except subprocess.CalledProcessError as fresh_exc:
-                        return subprocess.CompletedProcess(
-                            args=fresh_exc.cmd,
-                            returncode=fresh_exc.returncode,
-                            stdout=fresh_exc.stdout or "",
-                            stderr=fresh_exc.stderr or "",
-                        )
-            else:
-                try:
-                    result = run_codex_session(
-                        prompt,
-                        cwd=worktree_path,
-                        timeout=ci_driver_claude_timeout(),
-                        sandbox="workspace-write",
-                    )
-                except subprocess.CalledProcessError as exc:
-                    return subprocess.CompletedProcess(
-                        args=exc.cmd,
-                        returncode=exc.returncode,
-                        stdout=exc.stdout or "",
-                        stderr=exc.stderr or "",
-                    )
-            return subprocess.CompletedProcess(
-                args=[], returncode=0, stdout=result.stdout, stderr=result.stderr or ""
-            )
-
-        repo_slug = get_repo_slug(self.repo_root)
-        try:
-            stdout, _ = invoke_claude_with_session(
-                repo=repo_slug,
-                issue=issue_number,
-                agent=AGENT_CI_DRIVER,
-                prompt=prompt,
-                model=implementer_model(),
-                cwd=worktree_path,
-                timeout=ci_driver_claude_timeout(),
-                output_format="json",
-                allowed_tools="Read,Write,Edit,Glob,Grep,Bash",
-                extra_args=["--dangerously-skip-permissions"],
-                input_via_stdin=True,
-            )
-            return subprocess.CompletedProcess(args=[], returncode=0, stdout=stdout, stderr="")
-        except subprocess.CalledProcessError as exc:
-            return subprocess.CompletedProcess(
-                args=exc.cmd,
-                returncode=exc.returncode,
-                stdout=exc.stdout or "",
-                stderr=exc.stderr or "",
-            )
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.invoke_agent_session(
+            prompt=prompt,
+            session_id=session_id,
+            worktree_path=worktree_path,
+            issue_number=issue_number,
+            pr_number=pr_number,
+        )
 
     def _push_ci_fix(
         self,
@@ -2589,34 +1914,15 @@ class CIDriver:
         pr_head_branch: str,
         session_id: str | None,
     ) -> bool:
-        """Check head advancement, retry if needed, then push CI fixes.
-
-        Shared post-agent contract for both codex and claude providers (#846).
-        Returns True if fixes were pushed, False on any failure or no-commit.
-        """
-        if not self._head_advanced(worktree_path, pre_agent_sha, issue_number):  # noqa: SIM102
-            if not self._retry_no_commit_once(
-                issue_number=issue_number,
-                pr_number=pr_number,
-                worktree_path=worktree_path,
-                pr_head_branch=pr_head_branch,
-                pre_agent_sha=pre_agent_sha,
-                session_id=session_id,
-            ):
-                return False
-        if not self._ci_fix_head_is_pushable(worktree_path, issue_number):
-            return False
-        try:
-            push_current_branch_with_lease_on_divergence(
-                worktree_path,
-                branch=pr_head_branch,
-                push_ref=f"HEAD:{pr_head_branch}",
-            )
-            logger.info("Issue #%s: pushed CI fixes for PR #%s", issue_number, pr_number)
-            return True
-        except Exception as push_err:
-            logger.error("Issue #%s: git push failed after CI fix: %s", issue_number, push_err)
-            return False
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.push_ci_fix(
+            worktree_path=worktree_path,
+            pre_agent_sha=pre_agent_sha,
+            issue_number=issue_number,
+            pr_number=pr_number,
+            pr_head_branch=pr_head_branch,
+            session_id=session_id,
+        )
 
     def _retry_no_commit_once(
         self,
@@ -2629,123 +1935,16 @@ class CIDriver:
         session_id: str | None,
         max_retries: int = 2,
     ) -> bool:
-        """Re-invoke the agent up to ``max_retries`` times after a no-commit turn (#846).
-
-        A single no-op agent turn used to be terminal: with
-        ``max_fix_iterations=1`` the whole issue failed the instant the first
-        force-engagement retry also returned no commit. That cost real PRs
-        (AchaeanFleet #691/#683, Hermes #643). We now re-engage up to
-        ``max_retries`` times before recording the forensics marker, re-checking
-        between attempts so a PR that goes green (or where a commit lands)
-        short-circuits immediately.
-
-        Only fires when the PR still has failing required checks — a green PR
-        that arrived via concurrent activity should NOT be perturbed. Stays on
-        the same ``(repo, issue, AGENT_CI_DRIVER)`` session so the agent's
-        transcript is continuous; the existing PR and branch are preserved
-        (no new PR, no new branch, no force-push to a new ref).
-
-        Args:
-            issue_number: GitHub issue number for the PR.
-            pr_number: PR number to retry on.
-            worktree_path: Worktree the agent runs in (same as the first turn).
-            pr_head_branch: PR head branch name; the retry must not switch off it.
-            pre_agent_sha: HEAD SHA from before the first turn (used as the
-                no-commit baseline for the retry too — if the retry doesn't
-                advance past this, we treat it as repeated-no-commit).
-            session_id: Codex resume id (Claude resumes by deterministic UUID).
-
-        Returns:
-            True if the retry produced a new commit (caller should push).
-            False if CI is green now, the retry returned no commit again, or
-            any error path fired. On repeated-no-commit, writes a forensics
-            marker to ``state_dir``.
-
-        """
-        failing: list[str] = []
-        for retry in range(1, max_retries + 1):
-            failing = self._failing_required_check_names(pr_number)
-            dirty_tracked_changes = self._tracked_worktree_changes(worktree_path, issue_number)
-            if not failing and not dirty_tracked_changes:
-                logger.info(
-                    "Issue #%s: no-commit turn but PR #%s has no failing required checks "
-                    "and no tracked worktree changes; skipping force-engagement retry",
-                    issue_number,
-                    pr_number,
-                )
-                return False
-
-            review_threads_block = self._format_review_threads_block(pr_number)
-            retry_prompt = self._force_engagement_prompt(
-                issue_number=issue_number,
-                pr_number=pr_number,
-                worktree_path=worktree_path,
-                pr_head_branch=pr_head_branch,
-                failing_check_names=failing,
-                dirty_tracked_changes=dirty_tracked_changes,
-                review_threads_block=review_threads_block,
-            )
-
-            retry_reason = ", ".join(failing) if failing else "tracked worktree changes"
-            logger.warning(
-                "Issue #%s: no-commit on CI fix turn; re-invoking with "
-                "force-engagement prompt (retry %s/%s, reason: %s)",
-                issue_number,
-                retry,
-                max_retries,
-                retry_reason,
-            )
-
-            try:
-                retry_result = self._invoke_agent_session(
-                    prompt=retry_prompt,
-                    session_id=session_id,
-                    worktree_path=worktree_path,
-                    issue_number=issue_number,
-                    pr_number=pr_number,
-                )
-            except subprocess.TimeoutExpired:
-                logger.error(
-                    "Issue #%s: no-commit retry session timed out for PR #%s",
-                    issue_number,
-                    pr_number,
-                )
-                return False
-
-            if retry_result.returncode != 0:
-                logger.error(
-                    "Issue #%s: no-commit retry session failed for PR #%s: %s",
-                    issue_number,
-                    pr_number,
-                    (retry_result.stderr or "")[:300],
-                )
-                return False
-
-            if self._head_advanced(worktree_path, pre_agent_sha, issue_number):
-                return True
-
-            logger.warning(
-                "Issue #%s: still no commit on PR #%s after force-engagement retry %s/%s",
-                issue_number,
-                pr_number,
-                retry,
-                max_retries,
-            )
-
-        logger.error(
-            "Issue #%s: REPEATED no-commit on PR #%s after %s force-engagement "
-            "retries; marking and moving on",
-            issue_number,
-            pr_number,
-            max_retries,
-        )
-        self._record_repeated_no_commit(
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.retry_no_commit_once(
             issue_number=issue_number,
             pr_number=pr_number,
+            worktree_path=worktree_path,
             pr_head_branch=pr_head_branch,
-            failing_check_names=failing,
+            pre_agent_sha=pre_agent_sha,
+            session_id=session_id,
+            max_retries=max_retries,
         )
-        return False
 
     def _head_advanced(
         self,
@@ -2753,43 +1952,8 @@ class CIDriver:
         pre_agent_sha: str,
         issue_number: int,
     ) -> bool:
-        """Return True iff HEAD has moved past ``pre_agent_sha`` after the agent ran.
-
-        Called between the agent session and the push to detect the
-        no-commit-made case (#836). When ``pre_agent_sha`` still matches HEAD
-        the agent did not commit anything, so a force-with-lease push of
-        HEAD:<branch> would be a silent 0-exit no-op and the driver would
-        falsely log "pushed CI fixes". We instead log a warning and return
-        False so the iteration counts as failed.
-
-        Args:
-            worktree_path: Worktree to inspect.
-            pre_agent_sha: HEAD SHA captured right after the pre-agent sync.
-            issue_number: For log context.
-
-        Returns:
-            True if HEAD moved (something to push); False if it did not (or
-            if reading HEAD failed — we treat that as "don't push" too).
-
-        """
-        try:
-            post_agent_sha = run(["git", "rev-parse", "HEAD"], cwd=worktree_path).stdout.strip()
-        except subprocess.CalledProcessError as exc:
-            logger.error(
-                "Issue #%s: failed to read HEAD after CI fix session: %s",
-                issue_number,
-                (exc.stderr or exc.stdout or "")[:300],
-            )
-            return False
-        if post_agent_sha == pre_agent_sha:
-            logger.warning(
-                "Issue #%s: agent session produced no new commit (HEAD unchanged "
-                "at %s); skipping push and treating iteration as failed",
-                issue_number,
-                pre_agent_sha[:8],
-            )
-            return False
-        return True
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator._head_advanced(worktree_path, pre_agent_sha, issue_number)
 
     def _git_stdout_for_push_guard(
         self,
@@ -2798,26 +1962,10 @@ class CIDriver:
         argv: list[str],
         failure_message: str,
     ) -> str | None:
-        """Run a git inspection command for the CI pre-push guard."""
-        try:
-            result = run(argv, cwd=worktree_path, capture_output=True, check=False)
-        except subprocess.CalledProcessError as exc:
-            logger.error(
-                "Issue #%s: %s: %s",
-                issue_number,
-                failure_message,
-                (exc.stderr or exc.stdout or "")[:300],
-            )
-            return None
-        if result.returncode != 0:
-            logger.error(
-                "Issue #%s: %s: %s",
-                issue_number,
-                failure_message,
-                (result.stderr or result.stdout or "")[:300],
-            )
-            return None
-        return result.stdout or ""
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator._git_stdout_for_push_guard(
+            worktree_path, issue_number, argv, failure_message
+        )
 
     def _ci_fix_head_is_pushable(
         self,
@@ -2826,106 +1974,18 @@ class CIDriver:
         *,
         base_ref: str = "origin/main",
     ) -> bool:
-        """Return True when the post-agent worktree is safe to push.
-
-        ``_head_advanced`` only proves HEAD changed. A conflict-resolution agent
-        can still leave the index unmerged, leave tracked files uncommitted, or
-        accidentally detach at the base branch itself. None of those states may
-        be pushed to the PR head.
-        """
-        unmerged = self._git_stdout_for_push_guard(
-            worktree_path,
-            issue_number,
-            ["git", "diff", "--name-only", "--diff-filter=U"],
-            "failed to inspect merge state before push",
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator._ci_fix_head_is_pushable(
+            worktree_path, issue_number, base_ref=base_ref
         )
-        if unmerged is None:
-            return False
-        unmerged_paths = [line for line in unmerged.splitlines() if line.strip()]
-        if unmerged_paths:
-            logger.error(
-                "Issue #%s: refusing to push CI fix with unresolved merge paths: %s",
-                issue_number,
-                ", ".join(unmerged_paths[:10]),
-            )
-            return False
-
-        status = self._git_stdout_for_push_guard(
-            worktree_path,
-            issue_number,
-            ["git", "status", "--porcelain"],
-            "failed to inspect worktree status before push",
-        )
-        if status is None:
-            return False
-        tracked_dirty = [
-            line for line in status.splitlines() if line.strip() and not line.startswith("?? ")
-        ]
-        if tracked_dirty:
-            logger.error(
-                "Issue #%s: refusing to push CI fix with uncommitted tracked changes: %s",
-                issue_number,
-                ", ".join(tracked_dirty[:10]),
-            )
-            return False
-
-        ahead = self._git_stdout_for_push_guard(
-            worktree_path,
-            issue_number,
-            ["git", "rev-list", "--count", f"{base_ref}..HEAD"],
-            f"failed to inspect HEAD ahead of {base_ref} before push",
-        )
-        if ahead is None:
-            return False
-        try:
-            ahead_count = int(ahead.strip() or "0")
-        except ValueError:
-            logger.error(
-                "Issue #%s: refusing to push CI fix with invalid ahead count: %r",
-                issue_number,
-                ahead,
-            )
-            return False
-        if ahead_count <= 0:
-            logger.error(
-                "Issue #%s: refusing to push CI fix because HEAD has no commits ahead of %s",
-                issue_number,
-                base_ref,
-            )
-            return False
-        return True
 
     def _sync_worktree_and_snapshot_sha(
         self, issue_number: int, worktree_path: Path, pr_head_branch: str
     ) -> str | None:
-        """Sync worktree to remote PR head and snapshot HEAD SHA.
-
-        Returns the pre-agent SHA string, or ``None`` on any subprocess failure
-        (caller should return ``False`` immediately).
-
-        Syncing before the agent prevents the agent from committing on a stale
-        base and failing the force-with-lease push (#832). Snapshotting the SHA
-        lets the push helper detect sessions that return without committing (#836).
-        """
-        try:
-            sync_worktree_to_remote_branch(worktree_path, pr_head_branch)
-        except subprocess.CalledProcessError as exc:
-            logger.error(
-                "Issue #%s: failed to sync worktree to origin/%s before CI fix: %s",
-                issue_number,
-                pr_head_branch,
-                (exc.stderr or exc.stdout or "")[:300],
-            )
-            return None
-        try:
-            return run(["git", "rev-parse", "HEAD"], cwd=worktree_path).stdout.strip()
-        except subprocess.CalledProcessError as exc:
-            logger.error(
-                "Issue #%s: failed to snapshot HEAD before CI fix session: %s",
-                issue_number,
-                (exc.stderr or exc.stdout or "")[:300],
-            )
-            return None
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.sync_worktree_and_snapshot_sha(
+            issue_number, worktree_path, pr_head_branch
+        )
 
     def _build_ci_fix_prompt(
         self,
@@ -2936,27 +1996,14 @@ class CIDriver:
         pr_head_branch: str,
         advise_findings: str,
     ) -> str:
-        """Build the CI-fix agent prompt string."""
-        advise_block = ""
-        findings = advise_findings.strip()
-        if findings and not findings.startswith("<!-- advise step skipped"):
-            advise_block = f"## Prior Learnings from Team Knowledge Base\n\n{findings}\n\n---\n\n"
-        review_threads_block = self._format_review_threads_block(pr_number)
-        return (
-            f"{advise_block}{review_threads_block}"
-            f"Fix the CI failures for PR {pr_ref(pr_number)} (issue {issue_ref(issue_number)}).\n\n"
-            f"Working directory: {worktree_path}\n"
-            f"Current branch (DO NOT change): {pr_head_branch}\n\n"
-            f"CI failure logs:\n{ci_logs}\n\n"
-            "Fix the code to make the CI checks pass. After fixing:\n"
-            "1. Run: pixi run python -m pytest tests/ -v\n"
-            "2. Run: pre-commit run --all-files\n"
-            "3. Commit changes (do NOT push) on the current branch — DO NOT run "
-            "`git checkout -b`, `git switch -c`, or any other command that creates "
-            "or switches to a different branch\n"
-            "4. Every commit MUST be cryptographically signed (`git commit -S`); "
-            "NEVER use `--no-verify`.\n\n"
-            f"Commit message: fix: Address CI failures for PR {pr_ref(pr_number)}\n"
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.build_ci_fix_prompt(
+            issue_number,
+            pr_number,
+            worktree_path,
+            ci_logs,
+            pr_head_branch,
+            advise_findings,
         )
 
     def _run_ci_fix_session(
@@ -2970,74 +2017,15 @@ class CIDriver:
         *,
         pr_head_branch: str,
     ) -> bool:
-        """Invoke the selected coding agent to fix CI failures, then push the result.
-
-        Args:
-            issue_number: GitHub issue number.
-            pr_number: GitHub PR number.
-            worktree_path: Path to the checked-out worktree.
-            ci_logs: Combined CI failure log text.
-            session_id: Optional agent session ID to resume.
-            advise_findings: Prior learnings from the advise step, prepended to
-                the prompt as context. Empty or a skip marker contributes nothing.
-            pr_head_branch: The PR's head-branch name on the remote. The push
-                uses this as the destination refspec so the fix lands on the
-                actual PR branch even if the agent switched branches locally
-                during the session (#832).
-
-        Returns:
-            True if the fix session succeeded and the branch was pushed.
-
-        """
-        pre_agent_sha = self._sync_worktree_and_snapshot_sha(
-            issue_number, worktree_path, pr_head_branch
-        )
-        if pre_agent_sha is None:
-            return False
-
-        prompt = self._build_ci_fix_prompt(
+        """Delegate to CIFixOrchestrator (extracted #1357)."""
+        return self._fix_orchestrator.run_ci_fix_session(
             issue_number,
             pr_number,
             worktree_path,
             ci_logs,
-            pr_head_branch,
+            session_id,
             advise_findings,
-        )
-
-        try:
-            agent_result = self._invoke_agent_session(
-                prompt=prompt,
-                session_id=session_id,
-                worktree_path=worktree_path,
-                issue_number=issue_number,
-                pr_number=pr_number,
-            )
-        except subprocess.TimeoutExpired:
-            logger.error("Issue #%s: CI fix session timed out for PR #%s", issue_number, pr_number)
-            return False
-        except Exception as e:
-            logger.error(
-                "Issue #%s: CI fix session failed for PR #%s: %s", issue_number, pr_number, e
-            )
-            return False
-
-        if agent_result.returncode != 0:
-            logger.error(
-                "Issue #%s: CI fix session returned exit code %s: %s",
-                issue_number,
-                agent_result.returncode,
-                (agent_result.stderr or "")[:300],
-            )
-            return False
-
-        logger.debug("Issue #%s: CI fix output: %s", issue_number, agent_result.stdout[:500])
-        return self._push_ci_fix(
-            worktree_path=worktree_path,
-            pre_agent_sha=pre_agent_sha,
-            issue_number=issue_number,
-            pr_number=pr_number,
             pr_head_branch=pr_head_branch,
-            session_id=session_id,
         )
 
     def _enable_auto_merge(self, pr_number: int, is_bot_pr: bool = False) -> bool:
@@ -3115,121 +2103,12 @@ class CIDriver:
             return False
 
     def _run_drive_green_learnings(self, issue_number: int, pr_number: int) -> bool:
-        """Capture drive-green learnings under AGENT_CI_DRIVER (Session 3).
-
-        Runs after a PR reaches green and auto-merge is enabled, mirroring the
-        implementer's post-PR ``/learn`` step but scoped to *this* drive: what
-        made CI fail and how it was fixed. Resumes Session 3 (the
-        ``AGENT_CI_DRIVER`` session the fix session created) so the learnings
-        compound on the same transcript that did the work.
-
-        This is best-effort. Any failure is logged at WARNING and swallowed so
-        a flaky learnings step never flips a successful drive to failure.
-
-        Args:
-            issue_number: GitHub issue number.
-            pr_number: GitHub PR number.
-
-        Returns:
-            True if the learnings session completed, False otherwise.
-
-        """
-        prompt = build_learn_prompt(
-            f"You just drove PR {pr_ref(pr_number)} (issue {issue_ref(issue_number)}) "
-            "to green CI. Capture concise learnings about what made CI fail and how "
-            "you fixed it, scoped to this issue/PR."
-        )
-        self._last_drive_green_learn_evidence = mnemosyne_update_evidence("")
-        try:
-            repo_slug = get_repo_slug(self.repo_root)
-            # Best-effort: try to resume in the original worktree (so the
-            # AGENT_CI_DRIVER transcript is found by ``session_jsonl_path``).
-            # In the post-merge code path (#840) the PR's head branch may be
-            # gone from the remote, so ``_get_worktree_path`` could fail. In
-            # that case fall back to ``repo_root`` cwd — the prompt is
-            # self-contained, and a fresh ``--session-id`` create still
-            # captures the lesson.
-            try:
-                cwd = self._get_worktree_path(issue_number, pr_number)
-            except Exception as wt_err:
-                logger.info(
-                    "Issue #%s: no worktree available for /learn (%s); "
-                    "falling back to repo root for a fresh session",
-                    issue_number,
-                    wt_err,
-                )
-                cwd = self.repo_root
-            if is_codex(self.options.agent):
-                codex_result = run_codex_session(
-                    prompt,
-                    cwd=cwd,
-                    timeout=learn_claude_timeout(),
-                    sandbox="workspace-write",
-                )
-                self._last_drive_green_learn_evidence = mnemosyne_update_evidence(
-                    codex_result.stdout or ""
-                )
-                logger.info("Issue #%s: drive-green learnings captured with Codex", issue_number)
-                return True
-            stdout, _ = invoke_claude_with_session(
-                repo=repo_slug,
-                issue=issue_number,
-                agent=AGENT_CI_DRIVER,
-                prompt=prompt,
-                # /learn inherits the parent phase's model. drive-green resumes
-                # the implementer's session, and ``claude --resume`` is locked to
-                # the model that created it, so we must use implementer_model().
-                model=implementer_model(),
-                cwd=cwd,
-                timeout=learn_claude_timeout(),
-                output_format="text",
-                allowed_tools="Read,Write,Edit,Glob,Grep,Bash",
-                extra_args=["--dangerously-skip-permissions"],
-                input_via_stdin=True,
-            )
-            self._last_drive_green_learn_evidence = mnemosyne_update_evidence(stdout or "")
-            logger.info("Issue #%s: drive-green learnings captured", issue_number)
-            return True
-        except Exception as e:  # broad: external claude process; non-blocking
-            logger.warning(
-                "Issue #%s: drive-green learnings failed (non-fatal): %s",
-                issue_number,
-                e,
-            )
-            return False
+        """Delegate to PostMergeProcessor (extracted #1357)."""
+        return self._post_merge.run_drive_green_learnings(issue_number, pr_number)
 
     def _run_drive_green_compact(self, issue_number: int, pr_number: int) -> bool:
-        """Compact the AGENT_CI_DRIVER session transcript after /learn (#842).
-
-        Mirrors the cwd-derivation of ``_run_drive_green_learnings``: try the
-        worktree first so the deterministic JSONL probe in ``session_jsonl_path``
-        finds the transcript, fall back to ``repo_root`` when the branch is
-        already gone post-merge. Non-fatal.
-        """
-        if is_codex(self.options.agent):
-            logger.info(
-                "Issue #%s: skipping /compact (codex has no persisted "
-                "drive-green session to resume)",
-                issue_number,
-            )
-            return False
-        try:
-            cwd = self._get_worktree_path(issue_number, pr_number)
-        except Exception as wt_err:
-            logger.info(
-                "Issue #%s: no worktree available for /compact (%s); using repo root",
-                issue_number,
-                wt_err,
-            )
-            cwd = self.repo_root
-        repo_slug = get_repo_slug(self.repo_root)
-        return compact_session(
-            repo=repo_slug,
-            issue=issue_number,
-            agent=AGENT_CI_DRIVER,
-            cwd=cwd,
-            model=implementer_model(),
-        )
+        """Delegate to PostMergeProcessor (extracted #1357)."""
+        return self._post_merge.run_drive_green_compact(issue_number, pr_number)
 
     def _parse_json_block(self, text: str) -> dict[str, Any]:
         """Extract and parse the first JSON block from a text string.
