+++ pyproject.toml
@@ -241,7 +241,7 @@ omit = [
     "*/__init__.py",
     # Full orchestration loops require a live claude/gh CLI — integration test territory.
     # Pure-function helpers in these modules are tested in tests/unit/automation/ instead.
-    # The allowlist of these 12 modules is frozen in tests/unit/validation/test_omit_allowlist.py
+    # The allowlist of these 16 modules is frozen in tests/unit/validation/test_omit_allowlist.py
     # to prevent silent growth of the omit list, and their importability is guarded by
     # tests/integration/test_orchestration_smoke.py.
     "hephaestus/automation/implementer.py",
@@ -253,6 +253,13 @@ omit = [
     "hephaestus/automation/planner.py",
     "hephaestus/automation/address_review.py",
     "hephaestus/automation/ci_driver.py",
+    # CIDriver collaborators extracted from ci_driver.py (#1357 / refs #1179, #1289).
+    # Each drives a live gh/claude/codex session, so like ci_driver.py they are
+    # integration-only; their pure helpers are unit-tested in tests/unit/automation/.
+    "hephaestus/automation/pr_discovery.py",
+    "hephaestus/automation/ci_check_inspector.py",
+    "hephaestus/automation/ci_fix_orchestrator.py",
+    "hephaestus/automation/post_merge_processor.py",
     "hephaestus/automation/loop_runner.py",
     # curses_ui.py uses a live terminal; cannot be unit-tested without a real TTY
     "hephaestus/automation/curses_ui.py",
