+++ tests/unit/automation/test_ci_driver.py
@@ -272,17 +272,27 @@ def test_codex_ci_fix_session_falls_back_to_fresh_on_resume_failure(
     ]
 
     with (
-        patch("hephaestus.automation.ci_driver.resume_codex_session", side_effect=resume_error),
         patch(
-            "hephaestus.automation.ci_driver.run_codex_session",
+            "hephaestus.automation.ci_fix_orchestrator.resume_codex_session",
+            side_effect=resume_error,
+        ),
+        patch(
+            "hephaestus.automation.ci_fix_orchestrator.run_codex_session",
             return_value=fresh_result,
         ) as mock_fresh,
         patch(
-            "hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"
+            "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
         ) as mock_push,
-        patch("hephaestus.automation.ci_driver.sync_worktree_to_remote_branch") as mock_sync,
-        patch.object(driver, "_ci_fix_head_is_pushable", return_value=True),
-        patch("hephaestus.automation.ci_driver.run", side_effect=pre_post_sequence),
+        patch(
+            "hephaestus.automation.ci_fix_orchestrator.sync_worktree_to_remote_branch"
+        ) as mock_sync,
+        patch.object(driver._fix_orchestrator, "_ci_fix_head_is_pushable", return_value=True),
+        # The pre-agent SHA snapshot and the post-agent HEAD read (_head_advanced)
+        # both run inside ci_fix_orchestrator after the decomposition (#1357).
+        patch(
+            "hephaestus.automation.ci_fix_orchestrator.run",
+            side_effect=pre_post_sequence,
+        ),
     ):
         result = driver._run_ci_fix_session(
             issue_number=123,
@@ -323,15 +333,18 @@ def test_codex_ci_fix_session_skips_push_when_head_did_not_advance(
 
     with (
         patch(
-            "hephaestus.automation.ci_driver.run_codex_session",
+            "hephaestus.automation.ci_fix_orchestrator.run_codex_session",
             return_value=fresh_result,
         ),
         patch(
-            "hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"
+            "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
         ) as mock_push,
-        patch("hephaestus.automation.ci_driver.sync_worktree_to_remote_branch"),
+        patch("hephaestus.automation.ci_fix_orchestrator.sync_worktree_to_remote_branch"),
+        # Pre-agent SHA snapshot, post-agent HEAD read (_head_advanced) and the
+        # git-status read (_tracked_worktree_changes) all run inside
+        # ci_fix_orchestrator after the decomposition (#1357).
         patch(
-            "hephaestus.automation.ci_driver.run",
+            "hephaestus.automation.ci_fix_orchestrator.run",
             side_effect=[unchanged_sha, unchanged_sha, clean_status],
         ),
         patch(
@@ -339,7 +352,7 @@ def test_codex_ci_fix_session_skips_push_when_head_did_not_advance(
             return_value=[],
         ),
         patch(
-            "hephaestus.automation.ci_driver.gh_pr_checks",
+            "hephaestus.automation.ci_check_inspector.gh_pr_checks",
             return_value=[],
         ),
     ):
@@ -365,7 +378,7 @@ def test_ci_fix_head_not_pushable_with_unmerged_index(
 ) -> None:
     """A semantic conflict resolution is not pushable until the index is merged."""
     with patch(
-        "hephaestus.automation.ci_driver.run",
+        "hephaestus.automation.ci_fix_orchestrator.run",
         return_value=MagicMock(
             stdout="hephaestus/automation/loop_runner.py\n",
             stderr="",
@@ -385,7 +398,7 @@ def test_ci_fix_head_not_pushable_when_head_is_base_only(
         MagicMock(stdout="?? uv.lock\n", stderr="", returncode=0),  # generated untracked file
         MagicMock(stdout="0\n", stderr="", returncode=0),  # no commits ahead of origin/main
     ]
-    with patch("hephaestus.automation.ci_driver.run", side_effect=responses):
+    with patch("hephaestus.automation.ci_fix_orchestrator.run", side_effect=responses):
         assert driver._ci_fix_head_is_pushable(tmp_path, 993) is False
 
 
@@ -399,7 +412,7 @@ def test_ci_fix_head_pushable_with_clean_committed_pr_head(
         MagicMock(stdout="?? uv.lock\n", stderr="", returncode=0),
         MagicMock(stdout="1\n", stderr="", returncode=0),
     ]
-    with patch("hephaestus.automation.ci_driver.run", side_effect=responses):
+    with patch("hephaestus.automation.ci_fix_orchestrator.run", side_effect=responses):
         assert driver._ci_fix_head_is_pushable(tmp_path, 993) is True
 
 
@@ -496,8 +509,8 @@ class TestOpenPrsRemaining:
         driver.options.include_all_authors = True
         result_mock = MagicMock(stdout="[]")
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
-            patch("hephaestus.automation.ci_driver._gh_call", return_value=result_mock),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery._gh_call", return_value=result_mock),
         ):
             remaining = driver._list_open_prs_remaining()
         assert remaining == []
@@ -527,8 +540,8 @@ class TestOpenPrsRemaining:
         ]
         result_mock = MagicMock(stdout=json.dumps(rest_pulls))
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
-            patch("hephaestus.automation.ci_driver._gh_call", return_value=result_mock),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery._gh_call", return_value=result_mock),
             # #1328: merge-state is fetched per-PR via a separate gh pr view.
             patch.object(driver, "_pr_merge_state", side_effect=[("CLEAN", "MERGEABLE"), ("", "")]),
         ):
@@ -561,7 +574,9 @@ class TestOpenPrsRemaining:
         result_mock = MagicMock(
             stdout=json.dumps({"mergeStateStatus": "dirty", "mergeable": "conflicting"})
         )
-        with patch("hephaestus.automation.ci_driver._gh_call", return_value=result_mock) as mock_gh:
+        with patch(
+            "hephaestus.automation.pr_discovery._gh_call", return_value=result_mock
+        ) as mock_gh:
             merge_state, mergeable = driver._pr_merge_state(42)
         assert (merge_state, mergeable) == ("DIRTY", "CONFLICTING")
         cmd = mock_gh.call_args[0][0]
@@ -570,14 +585,14 @@ class TestOpenPrsRemaining:
 
     def test_pr_merge_state_unknown_marker_skips_query(self, driver: CIDriver) -> None:
         """#1328: the -1 unknown sentinel must not trigger a gh call."""
-        with patch("hephaestus.automation.ci_driver._gh_call") as mock_gh:
+        with patch("hephaestus.automation.pr_discovery._gh_call") as mock_gh:
             assert driver._pr_merge_state(-1) == ("", "")
         mock_gh.assert_not_called()
 
     def test_pr_merge_state_gh_failure_returns_unknown(self, driver: CIDriver) -> None:
         """#1328: a failed merge-state query degrades to unknown, never CONFLICTING."""
         with patch(
-            "hephaestus.automation.ci_driver._gh_call",
+            "hephaestus.automation.pr_discovery._gh_call",
             side_effect=subprocess.CalledProcessError(1, "gh"),
         ):
             assert driver._pr_merge_state(7) == ("", "")
@@ -590,9 +605,9 @@ class TestOpenPrsRemaining:
         # #821: this test verifies gh-failure handling, not author scope.
         driver.options.include_all_authors = True
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 side_effect=subprocess.CalledProcessError(1, "gh", stderr="rate limited"),
             ),
         ):
@@ -609,9 +624,9 @@ class TestOpenPrsRemaining:
         driver.options.include_all_authors = True
         result_mock = MagicMock(stdout="[]")
         with (
-            patch("hephaestus.automation.ci_driver.get_repo_info", return_value=("o", "r")),
+            patch("hephaestus.automation.pr_discovery.get_repo_info", return_value=("o", "r")),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=result_mock,
             ) as mock_gh,
         ):
@@ -1350,10 +1365,12 @@ class TestDriveGreenLearnings:
         with (
             patch.object(driver, "_get_worktree_path", return_value=tmp_path),
             patch(
-                "hephaestus.automation.ci_driver.run_codex_session",
+                "hephaestus.automation.post_merge_processor.run_codex_session",
                 return_value=mock_codex_result,
             ) as mock_codex,
-            patch("hephaestus.automation.ci_driver.invoke_claude_with_session") as mock_invoke,
+            patch(
+                "hephaestus.automation.post_merge_processor.invoke_claude_with_session"
+            ) as mock_invoke,
         ):
             result = driver._run_drive_green_learnings(123, 456)
 
@@ -1365,7 +1382,8 @@ class TestDriveGreenLearnings:
         assert "Only push skills to ProjectMnemosyne" in prompt
         assert mock_codex.call_args.kwargs["cwd"] == tmp_path
         mock_invoke.assert_not_called()
-        assert driver._last_drive_green_learn_evidence["mnemosyne_update_status"] == "confirmed"
+        # The learn-evidence cache moved into PostMergeProcessor (#1357).
+        assert driver._post_merge._last_learn_evidence["mnemosyne_update_status"] == "confirmed"
 
 
 # ---------------------------------------------------------------------------
@@ -1530,7 +1548,7 @@ class TestGetFailingCiLogs:
         driver.options.dry_run = False
         with (
             patch.object(driver, "_get_pr_branch", return_value="123-auto-impl"),
-            patch("hephaestus.automation.ci_driver._gh_call") as mock_gh,
+            patch("hephaestus.automation.ci_check_inspector._gh_call") as mock_gh,
         ):
             mock_gh.return_value = MagicMock(stdout="[]")
             driver._get_failing_ci_logs(pr_number=456)
@@ -1543,7 +1561,7 @@ class TestGetFailingCiLogs:
         """``gh run list`` must NOT be called without a ``--branch`` filter."""
         with (
             patch.object(driver, "_get_pr_branch", return_value="my-branch"),
-            patch("hephaestus.automation.ci_driver._gh_call") as mock_gh,
+            patch("hephaestus.automation.ci_check_inspector._gh_call") as mock_gh,
         ):
             mock_gh.return_value = MagicMock(stdout="[]")
             driver._get_failing_ci_logs(pr_number=10)
@@ -1653,7 +1671,7 @@ class TestFailingRequiredCheckNames:
 
     def test_all_green_returns_empty(self, driver: CIDriver) -> None:
         checks = [_make_check("lint", conclusion="success")]
-        with patch("hephaestus.automation.ci_driver.gh_pr_checks", return_value=checks):
+        with patch("hephaestus.automation.ci_check_inspector.gh_pr_checks", return_value=checks):
             assert driver._failing_required_check_names(pr_number=1) == []
 
     def test_one_required_failure_returned(self, driver: CIDriver) -> None:
@@ -1662,7 +1680,7 @@ class TestFailingRequiredCheckNames:
             _make_check("test", conclusion="failure"),
             _make_check("non-req", conclusion="failure", required=False),
         ]
-        with patch("hephaestus.automation.ci_driver.gh_pr_checks", return_value=checks):
+        with patch("hephaestus.automation.ci_check_inspector.gh_pr_checks", return_value=checks):
             names = driver._failing_required_check_names(pr_number=1)
         assert names == ["test"]
 
@@ -1671,13 +1689,13 @@ class TestFailingRequiredCheckNames:
         checks = [
             _make_check("only-check", conclusion="failure", required=False),
         ]
-        with patch("hephaestus.automation.ci_driver.gh_pr_checks", return_value=checks):
+        with patch("hephaestus.automation.ci_check_inspector.gh_pr_checks", return_value=checks):
             assert driver._failing_required_check_names(pr_number=1) == ["only-check"]
 
     def test_gh_failure_returns_empty(self, driver: CIDriver) -> None:
         """A gh blip must not promote the no-commit to a retry — empty list = skip."""
         with patch(
-            "hephaestus.automation.ci_driver.gh_pr_checks",
+            "hephaestus.automation.ci_check_inspector.gh_pr_checks",
             side_effect=RuntimeError("api down"),
         ):
             assert driver._failing_required_check_names(pr_number=1) == []
@@ -1757,16 +1775,18 @@ class TestNoCommitRetry:
         """No-commit + green CI + clean tracked tree → no retry."""
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="success")],
             ),
-            patch("hephaestus.automation.ci_driver.invoke_claude_with_session") as mock_invoke,
+            patch(
+                "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session"
+            ) as mock_invoke,
             patch(
                 "hephaestus.automation.ci_driver.gh_pr_list_unresolved_threads",
                 return_value=[],
             ),
             patch(
-                "hephaestus.automation.ci_driver.run",
+                "hephaestus.automation.ci_fix_orchestrator.run",
                 return_value=MagicMock(stdout="?? uv.lock\n", stderr="", returncode=0),
             ),
         ):
@@ -1795,7 +1815,7 @@ class TestNoCommitRetry:
         post_sha = MagicMock(stdout="deadbeef\n", stderr="", returncode=0)
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="success")],
             ),
             patch(
@@ -1803,10 +1823,10 @@ class TestNoCommitRetry:
                 return_value=[],
             ),
             patch(
-                "hephaestus.automation.ci_driver.invoke_claude_with_session",
+                "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
                 return_value=("done", "sess"),
             ) as mock_invoke,
-            patch("hephaestus.automation.ci_driver.run", side_effect=[status, post_sha]),
+            patch("hephaestus.automation.ci_fix_orchestrator.run", side_effect=[status, post_sha]),
         ):
             result = driver._retry_no_commit_once(
                 issue_number=993,
@@ -1833,7 +1853,7 @@ class TestNoCommitRetry:
         clean_status = MagicMock(stdout="", stderr="", returncode=0)
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="failure")],
             ),
             patch(
@@ -1841,10 +1861,13 @@ class TestNoCommitRetry:
                 return_value=[],
             ),
             patch(
-                "hephaestus.automation.ci_driver.invoke_claude_with_session",
+                "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
                 return_value=("done", "sess"),
             ) as mock_invoke,
-            patch("hephaestus.automation.ci_driver.run", side_effect=[clean_status, post_sha]),
+            patch(
+                "hephaestus.automation.ci_fix_orchestrator.run",
+                side_effect=[clean_status, post_sha],
+            ),
         ):
             result = driver._retry_no_commit_once(
                 issue_number=1,
@@ -1870,7 +1893,7 @@ class TestNoCommitRetry:
         clean_status = MagicMock(stdout="", stderr="", returncode=0)
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="failure")],
             ),
             patch(
@@ -1878,11 +1901,11 @@ class TestNoCommitRetry:
                 return_value=[],
             ),
             patch(
-                "hephaestus.automation.ci_driver.invoke_claude_with_session",
+                "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
                 return_value=("nope", "sess"),
             ) as mock_invoke,
             patch(
-                "hephaestus.automation.ci_driver.run",
+                "hephaestus.automation.ci_fix_orchestrator.run",
                 side_effect=[clean_status, unchanged, clean_status, unchanged],
             ),
         ):
@@ -1912,7 +1935,7 @@ class TestNoCommitRetry:
         """A subprocess error during retry → False, no marker (could not prove repeated)."""
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="failure")],
             ),
             patch(
@@ -1920,7 +1943,7 @@ class TestNoCommitRetry:
                 return_value=[],
             ),
             patch(
-                "hephaestus.automation.ci_driver.invoke_claude_with_session",
+                "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
                 side_effect=subprocess.CalledProcessError(1, ["claude"], stderr="boom"),
             ),
         ):
@@ -1944,7 +1967,7 @@ class TestNoCommitRetry:
         clean_status = MagicMock(stdout="", stderr="", returncode=0)
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="failure")],
             ),
             patch(
@@ -1952,11 +1975,14 @@ class TestNoCommitRetry:
                 return_value=[],
             ),
             patch(
-                "hephaestus.automation.ci_driver.resume_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.resume_codex_session",
                 return_value=AgentRunResult(stdout="ok", stderr="", session_id="s"),
             ) as mock_resume,
-            patch("hephaestus.automation.ci_driver.run_codex_session") as mock_fresh,
-            patch("hephaestus.automation.ci_driver.run", side_effect=[clean_status, post_sha]),
+            patch("hephaestus.automation.ci_fix_orchestrator.run_codex_session") as mock_fresh,
+            patch(
+                "hephaestus.automation.ci_fix_orchestrator.run",
+                side_effect=[clean_status, post_sha],
+            ),
         ):
             result = driver._retry_no_commit_once(
                 issue_number=1,
@@ -1984,11 +2010,11 @@ class TestBotPrDiscovery:
         driver.options.include_all_authors = True
         with (
             patch(
-                "hephaestus.automation.ci_driver.get_repo_info",
+                "hephaestus.automation.pr_discovery.get_repo_info",
                 return_value=("o", "r"),
             ),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout="[]"),
             ),
         ):
@@ -2004,11 +2030,11 @@ class TestBotPrDiscovery:
         ]
         with (
             patch(
-                "hephaestus.automation.ci_driver.get_repo_info",
+                "hephaestus.automation.pr_discovery.get_repo_info",
                 return_value=("o", "r"),
             ),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 return_value=MagicMock(stdout=json.dumps(raw)),
             ),
         ):
@@ -2021,11 +2047,11 @@ class TestBotPrDiscovery:
         driver.options.include_all_authors = True
         with (
             patch(
-                "hephaestus.automation.ci_driver.get_repo_info",
+                "hephaestus.automation.pr_discovery.get_repo_info",
                 return_value=("o", "r"),
             ),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 side_effect=subprocess.CalledProcessError(1, ["gh"], stderr="boom"),
             ),
         ):
@@ -2035,11 +2061,11 @@ class TestBotPrDiscovery:
         """Discovery returns empty dict when gh api times out (docstring contract)."""
         with (
             patch(
-                "hephaestus.automation.ci_driver.get_repo_info",
+                "hephaestus.automation.pr_discovery.get_repo_info",
                 return_value=("o", "r"),
             ),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 side_effect=subprocess.TimeoutExpired(cmd="gh", timeout=30),
             ),
         ):
@@ -2049,11 +2075,11 @@ class TestBotPrDiscovery:
         """Discovery returns empty dict when the gh binary is missing/unexecutable."""
         with (
             patch(
-                "hephaestus.automation.ci_driver.get_repo_info",
+                "hephaestus.automation.pr_discovery.get_repo_info",
                 return_value=("o", "r"),
             ),
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.pr_discovery._gh_call",
                 side_effect=FileNotFoundError(2, "No such file or directory", "gh"),
             ),
         ):
@@ -2366,16 +2392,18 @@ class TestMechanicalRebase:
         """A BEHIND PR rebases cleanly → pushes with lease, returns True."""
         with (
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.ci_fix_orchestrator._gh_call",
                 return_value=self._pr_state("BEHIND"),
             ),
             patch.object(driver, "_get_worktree_path", return_value=tmp_path),
-            patch("hephaestus.automation.ci_driver.sync_worktree_to_remote_branch") as mock_sync,
             patch(
-                "hephaestus.automation.ci_driver.rebase_worktree_onto", return_value=True
+                "hephaestus.automation.ci_fix_orchestrator.sync_worktree_to_remote_branch"
+            ) as mock_sync,
+            patch(
+                "hephaestus.automation.ci_fix_orchestrator.rebase_worktree_onto", return_value=True
             ) as mock_rebase,
             patch(
-                "hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"
+                "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
             ) as mock_push,
         ):
             result = driver._attempt_mechanical_rebase(
@@ -2391,14 +2419,16 @@ class TestMechanicalRebase:
         """A DIRTY PR whose rebase conflicts must NOT push — it returns False."""
         with (
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.ci_fix_orchestrator._gh_call",
                 return_value=self._pr_state("DIRTY"),
             ),
             patch.object(driver, "_get_worktree_path", return_value=tmp_path),
-            patch("hephaestus.automation.ci_driver.sync_worktree_to_remote_branch"),
-            patch("hephaestus.automation.ci_driver.rebase_worktree_onto", return_value=False),
+            patch("hephaestus.automation.ci_fix_orchestrator.sync_worktree_to_remote_branch"),
             patch(
-                "hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"
+                "hephaestus.automation.ci_fix_orchestrator.rebase_worktree_onto", return_value=False
+            ),
+            patch(
+                "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
             ) as mock_push,
         ):
             result = driver._attempt_mechanical_rebase(
@@ -2412,12 +2442,12 @@ class TestMechanicalRebase:
         """A CLEAN/BLOCKED PR is already on its base — no rebase, no push."""
         with (
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.ci_fix_orchestrator._gh_call",
                 return_value=self._pr_state("CLEAN"),
             ),
-            patch("hephaestus.automation.ci_driver.rebase_worktree_onto") as mock_rebase,
+            patch("hephaestus.automation.ci_fix_orchestrator.rebase_worktree_onto") as mock_rebase,
             patch(
-                "hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"
+                "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
             ) as mock_push,
         ):
             result = driver._attempt_mechanical_rebase(
@@ -2432,10 +2462,10 @@ class TestMechanicalRebase:
         """BLOCKED (green, waiting on review) is on-base — must not be rebased."""
         with (
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.ci_fix_orchestrator._gh_call",
                 return_value=self._pr_state("BLOCKED"),
             ),
-            patch("hephaestus.automation.ci_driver.rebase_worktree_onto") as mock_rebase,
+            patch("hephaestus.automation.ci_fix_orchestrator.rebase_worktree_onto") as mock_rebase,
         ):
             result = driver._attempt_mechanical_rebase(
                 issue_number=5, pr_number=50, acquired_slot=0
@@ -2448,15 +2478,17 @@ class TestMechanicalRebase:
         """The rebase targets the PR's actual baseRefName, not a hardcoded main."""
         with (
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.ci_fix_orchestrator._gh_call",
                 return_value=self._pr_state("BEHIND", base="develop"),
             ),
             patch.object(driver, "_get_worktree_path", return_value=tmp_path),
-            patch("hephaestus.automation.ci_driver.sync_worktree_to_remote_branch"),
+            patch("hephaestus.automation.ci_fix_orchestrator.sync_worktree_to_remote_branch"),
             patch(
-                "hephaestus.automation.ci_driver.rebase_worktree_onto", return_value=True
+                "hephaestus.automation.ci_fix_orchestrator.rebase_worktree_onto", return_value=True
             ) as mock_rebase,
-            patch("hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"),
+            patch(
+                "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
+            ),
         ):
             driver._attempt_mechanical_rebase(issue_number=5, pr_number=50, acquired_slot=0)
 
@@ -2468,10 +2500,10 @@ class TestMechanicalRebase:
         """A bad/empty gh response must be swallowed → False, never raise."""
         with (
             patch(
-                "hephaestus.automation.ci_driver._gh_call",
+                "hephaestus.automation.ci_fix_orchestrator._gh_call",
                 return_value=MagicMock(stdout="not json"),
             ),
-            patch("hephaestus.automation.ci_driver.rebase_worktree_onto") as mock_rebase,
+            patch("hephaestus.automation.ci_fix_orchestrator.rebase_worktree_onto") as mock_rebase,
         ):
             result = driver._attempt_mechanical_rebase(
                 issue_number=5, pr_number=50, acquired_slot=0
@@ -3130,7 +3162,7 @@ class TestRunDriveGreenCompact:
         self, driver: CIDriver, tmp_path: Path
     ) -> None:
         """Verify /compact runs exactly once and respects learn_captured_at gate."""
-        with patch("hephaestus.automation.ci_driver.compact_session") as mock_compact:
+        with patch("hephaestus.automation.post_merge_processor.compact_session") as mock_compact:
             mock_compact.return_value = True
 
             # First pass: should call compact_session
@@ -3145,7 +3177,7 @@ class TestRunDriveGreenCompact:
         self, driver: CIDriver, tmp_path: Path
     ) -> None:
         """Verify compact failure is non-fatal (returns False but doesn't raise)."""
-        with patch("hephaestus.automation.ci_driver.compact_session") as mock_compact:
+        with patch("hephaestus.automation.post_merge_processor.compact_session") as mock_compact:
             mock_compact.return_value = False
 
             # Should return False but not raise
@@ -3156,14 +3188,16 @@ class TestRunDriveGreenCompact:
         """Verify compact is skipped for codex (no persisted session)."""
         driver.options.agent = "codex"
 
-        with patch("hephaestus.automation.ci_driver.compact_session") as mock_compact:
+        with patch("hephaestus.automation.post_merge_processor.compact_session") as mock_compact:
             driver._run_drive_green_compact(842, 100)
             mock_compact.assert_not_called()
 
     def test_drive_green_compact_uses_worktree_path(self, driver: CIDriver, tmp_path: Path) -> None:
         """Verify compact_session is called with the worktree path."""
         with patch.object(driver, "_get_worktree_path", return_value=tmp_path):
-            with patch("hephaestus.automation.ci_driver.compact_session") as mock_compact:
+            with patch(
+                "hephaestus.automation.post_merge_processor.compact_session"
+            ) as mock_compact:
                 mock_compact.return_value = True
 
                 driver._run_drive_green_compact(842, 100)
@@ -3177,7 +3211,9 @@ class TestRunDriveGreenCompact:
     ) -> None:
         """Verify compact_session uses repo_root when worktree is not available."""
         with patch.object(driver, "_get_worktree_path", side_effect=RuntimeError("No worktree")):
-            with patch("hephaestus.automation.ci_driver.compact_session") as mock_compact:
+            with patch(
+                "hephaestus.automation.post_merge_processor.compact_session"
+            ) as mock_compact:
                 mock_compact.return_value = True
 
                 driver._run_drive_green_compact(842, 100)
@@ -3256,7 +3292,7 @@ class TestInvokeAgentSession:
 
     def test_claude_success_returns_rc0(self, driver: CIDriver, tmp_path: Path) -> None:
         with patch(
-            "hephaestus.automation.ci_driver.invoke_claude_with_session",
+            "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
             return_value=("output text", "sess-id"),
         ) as mock_invoke:
             result = driver._invoke_agent_session(
@@ -3272,7 +3308,7 @@ class TestInvokeAgentSession:
 
     def test_claude_error_returns_nonzero_rc(self, driver: CIDriver, tmp_path: Path) -> None:
         with patch(
-            "hephaestus.automation.ci_driver.invoke_claude_with_session",
+            "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
             side_effect=subprocess.CalledProcessError(1, ["claude"], stderr="boom"),
         ):
             result = driver._invoke_agent_session(
@@ -3289,10 +3325,10 @@ class TestInvokeAgentSession:
         driver.options.agent = "codex"
         with (
             patch(
-                "hephaestus.automation.ci_driver.resume_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.resume_codex_session",
                 return_value=AgentRunResult(stdout="ok", stderr="", session_id="s"),
             ) as mock_resume,
-            patch("hephaestus.automation.ci_driver.run_codex_session") as mock_fresh,
+            patch("hephaestus.automation.ci_fix_orchestrator.run_codex_session") as mock_fresh,
         ):
             result = driver._invoke_agent_session(
                 prompt="fix it",
@@ -3311,11 +3347,11 @@ class TestInvokeAgentSession:
         driver.options.agent = "codex"
         with (
             patch(
-                "hephaestus.automation.ci_driver.resume_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.resume_codex_session",
                 side_effect=subprocess.CalledProcessError(1, ["codex"], stderr="resume-fail"),
             ),
             patch(
-                "hephaestus.automation.ci_driver.run_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.run_codex_session",
                 return_value=AgentRunResult(stdout="fresh ok", stderr="", session_id="s2"),
             ) as mock_fresh,
         ):
@@ -3336,11 +3372,11 @@ class TestInvokeAgentSession:
         driver.options.agent = "codex"
         with (
             patch(
-                "hephaestus.automation.ci_driver.resume_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.resume_codex_session",
                 side_effect=subprocess.CalledProcessError(1, ["codex"], stderr="resume-fail"),
             ),
             patch(
-                "hephaestus.automation.ci_driver.run_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.run_codex_session",
                 side_effect=subprocess.CalledProcessError(2, ["codex"], stderr="fresh-fail"),
             ),
         ):
@@ -3360,7 +3396,7 @@ class TestInvokeAgentSession:
         """No session_id + fresh codex fails → CompletedProcess(returncode!=0), no exception."""
         driver.options.agent = "codex"
         with patch(
-            "hephaestus.automation.ci_driver.run_codex_session",
+            "hephaestus.automation.ci_fix_orchestrator.run_codex_session",
             side_effect=subprocess.CalledProcessError(3, ["codex"], stderr="fail"),
         ):
             result = driver._invoke_agent_session(
@@ -3375,9 +3411,9 @@ class TestInvokeAgentSession:
     def test_codex_no_session_runs_fresh(self, driver: CIDriver, tmp_path: Path) -> None:
         driver.options.agent = "codex"
         with (
-            patch("hephaestus.automation.ci_driver.resume_codex_session") as mock_resume,
+            patch("hephaestus.automation.ci_fix_orchestrator.resume_codex_session") as mock_resume,
             patch(
-                "hephaestus.automation.ci_driver.run_codex_session",
+                "hephaestus.automation.ci_fix_orchestrator.run_codex_session",
                 return_value=AgentRunResult(stdout="done", stderr="", session_id="s3"),
             ) as mock_fresh,
         ):
@@ -3395,7 +3431,7 @@ class TestInvokeAgentSession:
     def test_timeout_propagates_to_caller(self, driver: CIDriver, tmp_path: Path) -> None:
         """TimeoutExpired escapes the helper so callers can log a distinct timeout message."""
         with patch(
-            "hephaestus.automation.ci_driver.invoke_claude_with_session",
+            "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
             side_effect=subprocess.TimeoutExpired(cmd=["claude"], timeout=60),
         ):
             with pytest.raises(subprocess.TimeoutExpired):
@@ -3426,11 +3462,11 @@ class TestPushCiFix:
         no_untracked = MagicMock(stdout="", returncode=0)
         with (
             patch(
-                "hephaestus.automation.ci_driver.run",
+                "hephaestus.automation.ci_fix_orchestrator.run",
                 side_effect=[post_sha, clean_status, no_untracked, ahead_count],
             ),
             patch(
-                "hephaestus.automation.ci_driver.push_current_branch_with_lease_on_divergence"
+                "hephaestus.automation.ci_fix_orchestrator.push_current_branch_with_lease_on_divergence"
             ) as mock_push,
         ):
             result = driver._push_ci_fix(
@@ -3451,7 +3487,7 @@ class TestPushCiFix:
         clean_status = MagicMock(stdout="", stderr="", returncode=0)
         with (
             patch(
-                "hephaestus.automation.ci_driver.gh_pr_checks",
+                "hephaestus.automation.ci_check_inspector.gh_pr_checks",
                 return_value=[_make_check("lint", conclusion="failure")],
             ),
             patch(
@@ -3460,11 +3496,11 @@ class TestPushCiFix:
             ),
             # Agent fails → _retry_no_commit_once returns False immediately
             patch(
-                "hephaestus.automation.ci_driver.invoke_claude_with_session",
+                "hephaestus.automation.ci_fix_orchestrator.invoke_claude_with_session",
                 side_effect=subprocess.CalledProcessError(1, ["claude"], stderr="err"),
             ),
             patch(
-                "hephaestus.automation.ci_driver.run",
+                "hephaestus.automation.ci_fix_orchestrator.run",
                 side_effect=[unchanged, clean_status],
             ),
         ):
@@ -3483,7 +3519,7 @@ class TestPushCiFix:
         # Unmerged index entries → not pushable
         unmerged = MagicMock(stdout="UU conflict.py\n", stderr="", returncode=0)
         with patch(
-            "hephaestus.automation.ci_driver.run",
+            "hephaestus.automation.ci_fix_orchestrator.run",
             side_effect=[post_sha, unmerged],
         ):
             result = driver._push_ci_fix(
